backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
Flask-Limiter
Flask-WTF
Werkzeug
sentry-sdk[flask]
requests
python-dotenv
Flask-Cors
email_validator
psycopg2-binary
supabase
pytest
gunicorn
PyJWT
cryptography
pytest-mock
freezegun
flask-marshmallow
marshmallow-sqlalchemy
redis
Flask-Caching
flask-smorest
geopy
ImageHash
google-cloud-vision
Pillow
Flask-Migrate
